@@ -49,19 +49,19 @@
 	},
 	"dependencies": {
 		"@eslint-community/eslint-plugin-eslint-comments": "^4.4.1",
-		"@eslint-react/eslint-plugin": "^1.30.1",
-		"@html-eslint/eslint-plugin": "^0.35.0",
-		"@html-eslint/parser": "^0.35.0",
+		"@eslint-react/eslint-plugin": "^1.31.0",
+		"@html-eslint/eslint-plugin": "^0.35.2",
+		"@html-eslint/parser": "^0.35.2",
 		"@pinojs/json-colorizer": "^4.0.0",
 		"@typescript-eslint/eslint-plugin": "^8.26.0",
 		"@typescript-eslint/parser": "^8.26.0",
 		"@vitest/eslint-plugin": "^1.1.36",
 		"astro-eslint-parser": "^1.2.1",
-		"eslint": "^9.21.0",
+		"eslint": "^9.22.0",
 		"eslint-config-flat-gitignore": "^2.1.0",
 		"eslint-flat-config-utils": "^2.0.1",
 		"eslint-import-resolver-typescript": "^3.8.3",
-		"eslint-mdx": "^3.1.5",
+		"eslint-mdx": "^3.2.0",
 		"eslint-plugin-astro": "^1.3.1",
 		"eslint-plugin-depend": "^0.12.0",
 		"eslint-plugin-html": "^8.1.2",
@@ -69,13 +69,13 @@
 		"eslint-plugin-jsdoc": "^50.6.3",
 		"eslint-plugin-jsonc": "^2.19.1",
 		"eslint-plugin-jsx-a11y": "^6.10.2",
-		"eslint-plugin-mdx": "^3.1.5",
-		"eslint-plugin-n": "^17.16.1",
+		"eslint-plugin-mdx": "^3.2.0",
+		"eslint-plugin-n": "^17.16.2",
 		"eslint-plugin-no-only-tests": "^3.3.0",
-		"eslint-plugin-package-json": "^0.26.2",
-		"eslint-plugin-perfectionist": "^4.9.0",
+		"eslint-plugin-package-json": "^0.26.3",
+		"eslint-plugin-perfectionist": "^4.10.1",
 		"eslint-plugin-regexp": "^2.7.0",
-		"eslint-plugin-svelte": "^3.0.2",
+		"eslint-plugin-svelte": "^3.0.3",
 		"eslint-plugin-toml": "^0.12.0",
 		"eslint-plugin-unicorn": "^57.0.0",
 		"eslint-plugin-yml": "^1.17.0",
@@ -86,8 +86,8 @@
 		"jsonc-eslint-parser": "^2.4.0",
 		"local-pkg": "^1.1.1",
 		"prettier": "^3.5.3",
-		"sort-package-json": "^2.15.1",
-		"svelte-eslint-parser": "^1.0.0",
+		"sort-package-json": "^3.0.0",
+		"svelte-eslint-parser": "^1.0.1",
 		"toml-eslint-parser": "^0.10.0",
 		"yaml-eslint-parser": "^1.3.0"
 	},
@@ -95,11 +95,11 @@
 		"@types/eslint-config-prettier": "^6.11.3",
 		"@types/eslint-plugin-jsx-a11y": "^6.10.0",
 		"dot-prop": "^9.0.0",
-		"eslint-config-prettier": "^10.0.2",
+		"eslint-config-prettier": "^10.1.1",
 		"eslint-config-xo-typescript": "^7.0.0",
 		"eslint-typegen": "^2.0.0",
 		"globby": "^14.1.0",
-		"svelte": "^5.22.1",
+		"svelte": "^5.22.6",
 		"tsup": "^8.4.0"
 	},
 	"engines": {
